loading component added, files updated
