+++ src/components/LoadingState.css
@@ -1,34 +1,15 @@
-/* .carousel-wrapper {
-    max-width: 1200px;
-    width: 100%;
-    margin: 0 auto;
-    padding: 0 1rem;
-    box-sizing: border-box;
-} */
- 
 .skeleton__loading {
-  display: flex;
-  flex-direction: column;
-  gap: 1.5rem;
-  padding: 1rem;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
 }
 
-.skeleton-title {
+.skeleton-tiles-grid {
   display: flex;
-  height: 40px;
-  width: 25%;
-  margin: 0 auto;
-  border-radius: 6px;
-  background: linear-gradient(90deg, #333 25%, #444 50%, #333 75%);
-  background-size: 200% 100%;
-  animation: shimmer 1.5s infinite;
-}
-
-.skeleton-tiles {
-  display: flex;
+  align-items: center;
+  justify-content: center;
   gap: 1.5rem;
-  margin: 0 124px;
-  padding: 0 1rem;
 }
 
 .skeleton__tile {
@@ -36,28 +17,31 @@
   flex-direction: column;
   align-items: center;
   gap: 0.75rem;
-  width: 100%;
-  border: 2px solid #555;
+  border: 2px solid #777;
   border-radius: 12px;
   padding: 1rem;
+  background-color: #777;
+  min-height: 250px; 
+  min-width: 250px;
   box-sizing: border-box;
-  background-color: #555;
 }
+
 
 .skeleton__tile-img {
   width: 100%;
-  height: 120px;
+  height: 150px;
   border-radius: 10px;
-  background: linear-gradient(90deg, #333 25%, #444 50%, #333 75%);
+  background: linear-gradient(90deg, #aaa 25%, #bbb 50%, #aaa 75%);
   background-size: 200% 100%;
   animation: shimmer 1.5s infinite;
 }
+
 
 .skeleton__tile-author {
   width: 50px;
   height: 50px;
   border-radius: 50%;
-  background: linear-gradient(90deg, #333 25%, #444 50%, #333 75%);
+  background: linear-gradient(90deg, #aaa 25%, #bbb 50%, #aaa 75%);
   background-size: 200% 100%;
   animation: shimmer 1.5s infinite;
 }
@@ -66,7 +50,7 @@
 .skeleton__tile-code {
   height: 12px;
   border-radius: 6px;
-  background: linear-gradient(90deg, #333 25%, #444 50%, #333 75%);
+  background: linear-gradient(90deg, #aaa 25%, #bbb 50%, #aaa 75%);
   background-size: 200% 100%;
   animation: shimmer 1.5s infinite;
 }
@@ -87,3 +71,4 @@
     background-position: 200% 0;
   }
 }
+
