+++ src/components/home/NewItems.jsx
@@ -4,17 +4,14 @@ import axios from "axios";
 import "slick-carousel/slick/slick.css";
 import "slick-carousel/slick/slick-theme.css";
 import Slider from "react-slick";
-import "./HotCollections.css";
 import Countdown from "../Countdown";
+import LoadingState from "../LoadingState";
 
 const NewItems = () => {
   const [newItems, setNewItems] = useState([]);
-  const [loading, setLoading] = useState(false);
-  const fullArray = [1, 2, 3, 4];
-  const [visibleItems, setVisibleItems] = useState(fullArray);
+  const [loading, setLoading] = useState(true);
 
   async function getNewItems() {
-    setLoading(true);
     const { data } = await axios.get(
       `https://us-central1-nft-cloud-functions.cloudfunctions.net/newItems`
     );
@@ -24,21 +21,6 @@ const NewItems = () => {
 
   useEffect(() => {
     getNewItems();
-  }, []);
-
-  useEffect(() => {
-    const updateVisibleItems = () => {
-      const width = window.innerWidth;
-
-      if (width >= 1200) setVisibleItems(fullArray.slice(0, 4));
-      else if (width >= 768) setVisibleItems(fullArray.slice(0, 3));
-      else if (width >= 480) setVisibleItems(fullArray.slice(0, 2));
-      else setVisibleItems(fullArray.slice(0, 1));
-    };
-    updateVisibleItems();
-    window.addEventListener("resize", updateVisibleItems);
-
-    return () => window.removeEventListener("resize", updateVisibleItems);
   }, []);
 
   var settings = {
@@ -71,22 +53,7 @@ const NewItems = () => {
 
   return (
     <section id="section-items" className="no-bottom">
-      {loading ? (
-        <div className="skeleton__loading">
-          <div className="skeleton-title"></div>
-          <div className="small-border bg-color-2"></div>
-          <div className="skeleton-tiles">
-            {visibleItems.map((item, i) => (
-              <div className="skeleton__tile" key={i}>
-                <div className="skeleton__tile-img"></div>
-                <div className="skeleton__tile-author"></div>
-                <div className="skeleton__tile-name"></div>
-                <div className="skeleton__tile-code"></div>
-              </div>
-            ))}
-          </div>
-        </div>
-      ) : (
+      <LoadingState loading={loading}>
         <div className="container">
           <div className="col-lg-12">
             <div className="text-center">
@@ -179,7 +146,7 @@ const NewItems = () => {
           opacity: 1 !important; 
         }`}</style>
         </div>
-      )}
+      </LoadingState>
     </section>
   );
 };
